+++ packages/snaps-rpc-methods/src/restricted/getBip32Entropy.test.ts
@@ -79,7 +79,7 @@ describe('getBip32EntropyImplementation', () => {
       async (_selector, operation) =>
         operation({
           keyring: {
-            type: 'HD Key Tree',
+            type: 'hd',
             mnemonic: TEST_SECRET_RECOVERY_PHRASE_BYTES,
             seed: TEST_SECRET_RECOVERY_PHRASE_SEED_BYTES,
           },
@@ -345,7 +345,7 @@ describe('getBip32EntropyImplementation', () => {
         async (_selector, operation) =>
           operation({
             keyring: {
-              type: 'HD Key Tree',
+              type: 'hd',
             },
           }),
       );
@@ -375,7 +375,7 @@ describe('getBip32EntropyImplementation', () => {
         async (_selector, operation) =>
           operation({
             keyring: {
-              type: 'HD Key Tree',
+              type: 'hd',
             },
           }),
       );
@@ -405,7 +405,7 @@ describe('getBip32EntropyImplementation', () => {
         async (_selector, operation) =>
           operation({
             keyring: {
-              type: 'HD Key Tree',
+              type: 'hd',
             },
           }),
       );
@@ -439,7 +439,7 @@ describe('getBip32EntropyImplementation', () => {
         async (_selector, operation) =>
           operation({
             keyring: {
-              type: 'HD Key Tree',
+              type: 'hd',
             },
           }),
       );