+++ packages/snaps-rpc-methods/src/restricted/getBip32PublicKey.test.ts
@@ -79,7 +79,7 @@ describe('getBip32PublicKeyImplementation', () => {
       async (_selector, operation) =>
         operation({
           keyring: {
-            type: 'HD Key Tree',
+            type: 'hd',
             mnemonic: TEST_SECRET_RECOVERY_PHRASE_BYTES,
             seed: TEST_SECRET_RECOVERY_PHRASE_SEED_BYTES,
           },