+++ packages/snaps-rpc-methods/src/restricted/getBip44Entropy.test.ts
@@ -78,7 +78,7 @@ describe('getBip44EntropyImplementation', () => {
       async (_selector, operation) =>
         operation({
           keyring: {
-            type: 'HD Key Tree',
+            type: 'hd',
             seed: TEST_SECRET_RECOVERY_PHRASE_SEED_BYTES,
           },
         }),