+++ packages/snaps-rpc-methods/src/restricted/getEntropy.test.ts
@@ -59,7 +59,7 @@ describe('getEntropyImplementation', () => {
       async (_selector, operation) =>
         operation({
           keyring: {
-            type: 'HD Key Tree',
+            type: 'hd',
             seed: TEST_SECRET_RECOVERY_PHRASE_SEED_BYTES,
           },
         }),